requirements.txt
asgiref==3.7.2
Django==4.2.2
django-throttle-requests==0.7.0
sqlparse==0.4.4
